Make minimum required version C++20, but '23 is still preferable. Fix f64 = double instead of long double, that's a very old bad idea I never fixed. Also patch in SettingWrapper so TOML won't complain in linux (since a long double there is 16 bytes instead of the expected 8).

+++ src/Assistants/SettingWrapper.hpp
@@ -19,7 +19,7 @@
 using PtrVariant = std::variant<
 	std::int8_t  *, std::int16_t  *, std::int32_t  *, std::int64_t  *,
 	std::uint8_t *, std::uint16_t *, std::uint32_t *,
-	bool *, float *, long double *, std::string *
+	bool *, float *, double *, std::string *
 >;
 
 template<typename V, typename Variant>
